Refactor in preparation for importing data from already-published library
