+++ apps/src/code-studio/components/libraries/LibraryCreationDialog.jsx
@@ -4,38 +4,18 @@ import PropTypes from 'prop-types';
 import Dialog, {Body} from '@cdo/apps/templates/Dialog';
 import {connect} from 'react-redux';
 import {hideLibraryCreationDialog} from '../shareDialogRedux';
-import libraryParser from './libraryParser';
 import i18n from '@cdo/locale';
 import PadAndCenter from '@cdo/apps/templates/teacherDashboard/PadAndCenter';
 import {Heading1, Heading2} from '@cdo/apps/lib/ui/Headings';
-import annotationList from '@cdo/apps/acemode/annotationList';
 import Spinner from '../../pd/components/spinner';
 import Button from '@cdo/apps/templates/Button';
-import color from '@cdo/apps/util/color';
+import LibraryPublisher from './LibraryPublisher';
+import libraryLoader from './libraryLoader';
 
 const styles = {
-  alert: {
-    color: color.red,
-    width: '90%',
-    paddingTop: 8
-  },
   libraryBoundary: {
     padding: 10,
     width: '90%'
-  },
-  largerCheckbox: {
-    width: 20,
-    height: 20,
-    marginLeft: 0,
-    marginRight: 10,
-    marginTop: 10,
-    marginBottom: 10
-  },
-  functionItem: {
-    marginBottom: 20
-  },
-  textarea: {
-    width: 400
   },
   centerContent: {
     display: 'flex',
@@ -49,19 +29,6 @@ const styles = {
   button: {
     marginLeft: 10,
     marginRight: 10
-  },
-  info: {
-    fontSize: 12,
-    fontStyle: 'italic'
-  },
-  textInput: {
-    fontSize: 14,
-    padding: 6,
-    color: color.dimgray
-  },
-  description: {
-    width: '98%',
-    resize: 'vertical'
   }
 };
 
@@ -69,14 +36,12 @@ const styles = {
  * @readonly
  * @enum {string}
  */
-export const PublishState = {
+export const DialogState = {
   LOADING: 'loading',
   DONE_LOADING: 'done_loading',
   PUBLISHED: 'published',
-  ERROR_PUBLISH: 'error_publish',
   CODE_ERROR: 'code_error',
-  NO_FUNCTIONS: 'no_functions',
-  INVALID_INPUT: 'invalid_input'
+  NO_FUNCTIONS: 'no_functions'
 };
 
 class LibraryCreationDialog extends React.Component {
@@ -87,13 +52,9 @@ class LibraryCreationDialog extends React.Component {
   };
 
   state = {
-    librarySource: '',
-    sourceFunctionList: [],
-    publishState: PublishState.LOADING,
+    dialogState: DialogState.LOADING,
     libraryName: '',
-    libraryDescription: '',
-    canPublish: false,
-    selectedFunctions: {}
+    libraryDetails: {}
   };
 
   componentDidUpdate(prevProps) {
@@ -103,85 +64,25 @@ class LibraryCreationDialog extends React.Component {
   }
 
   onOpen = () => {
-    var error = annotationList.getJSLintAnnotations().find(annotation => {
-      return annotation.type === 'error';
-    });
-    if (error) {
-      this.setState({publishState: PublishState.CODE_ERROR});
-      return;
-    }
-
-    dashboard.project.getUpdatedSourceAndHtml_(response => {
-      let functionsList = libraryParser.getFunctions(response.source);
-      if (!functionsList || functionsList.length === 0) {
-        this.setState({publishState: PublishState.NO_FUNCTIONS});
-        return;
-      }
-      let librarySource = response.source;
-      if (response.libraries) {
-        response.libraries.forEach(library => {
-          librarySource =
-            libraryParser.createLibraryClosure(library) + librarySource;
-        });
-      }
-      this.setState({
-        libraryName: libraryParser.suggestName(
-          dashboard.project.getLevelName()
-        ),
-        librarySource: librarySource,
-        publishState: PublishState.DONE_LOADING,
-        sourceFunctionList: functionsList
-      });
-    });
+    libraryLoader.load(
+      () => this.setState({dialogState: DialogState.CODE_ERROR}),
+      () => this.setState({dialogState: DialogState.NO_FUNCTIONS}),
+      libraryDetails =>
+        this.setState({
+          dialogState: DialogState.DONE_LOADING,
+          libraryDetails: libraryDetails
+        })
+    );
   };
 
   handleClose = () => {
-    this.setState({publishState: PublishState.LOADING});
+    this.setState({dialogState: DialogState.LOADING});
     this.props.onClose();
   };
 
   copyChannelId = () => {
     this.channelId.select();
     document.execCommand('copy');
-  };
-
-  publish = () => {
-    let {
-      librarySource,
-      libraryName,
-      libraryDescription,
-      selectedFunctions,
-      sourceFunctionList
-    } = this.state;
-    let functionsToPublish = sourceFunctionList.filter(sourceFunction => {
-      return selectedFunctions[sourceFunction.functionName];
-    });
-
-    if (!(libraryDescription && functionsToPublish.length > 0)) {
-      this.setState({publishState: PublishState.INVALID_INPUT});
-      return;
-    }
-
-    let libraryJson = libraryParser.createLibraryJson(
-      librarySource,
-      functionsToPublish,
-      libraryName,
-      libraryDescription
-    );
-
-    // TODO: Display final version of error and success messages to the user.
-    this.props.clientApi.publish(
-      libraryJson,
-      error => {
-        console.warn(`Error publishing library: ${error}`);
-        this.setState({publishState: PublishState.ERROR_PUBLISH});
-      },
-      () => {
-        this.setState({publishState: PublishState.PUBLISHED});
-      }
-    );
-    dashboard.project.setLibraryName(this.state.libraryName);
-    dashboard.project.setLibraryDescription(libraryDescription);
   };
 
   displayError = errorMessage => {
@@ -194,14 +95,6 @@ class LibraryCreationDialog extends React.Component {
         <Spinner />
       </div>
     );
-  };
-
-  setLibraryName = event => {
-    let sanitizedName = libraryParser.sanitizeName(event.target.value);
-    if (sanitizedName === this.state.libraryName) {
-      return;
-    }
-    this.setState({libraryName: sanitizedName});
   };
 
   displaySuccess = () => {
@@ -233,130 +126,34 @@ class LibraryCreationDialog extends React.Component {
     );
   };
 
-  displayNameInput = () => {
-    return (
-      <div>
-        <input
-          style={styles.textInput}
-          type="text"
-          value={this.state.libraryName}
-          onChange={this.setLibraryName}
-          onBlur={event =>
-            this.setState({
-              libraryName: libraryParser.suggestName(event.target.value)
-            })
-          }
-        />
-        <div style={styles.info}>{i18n.libraryNameRequirements()}</div>
-      </div>
-    );
-  };
-
-  resetErrorMessage = () => {
-    if (
-      this.state.libraryDescription &&
-      Object.values(this.state.selectedFunctions).find(value => value) &&
-      this.state.publishState === PublishState.INVALID_INPUT
-    ) {
-      this.setState({publishState: PublishState.DONE_LOADING});
-    }
-  };
-
-  displayDescription = () => {
-    return (
-      <textarea
-        rows="2"
-        cols="200"
-        style={{...styles.textInput, ...styles.description}}
-        placeholder={i18n.libraryDescriptionPlaceholder()}
-        value={this.state.libraryDescription}
-        onChange={event =>
-          this.setState({libraryDescription: event.target.value})
-        }
-        onBlur={this.resetErrorMessage}
-      />
-    );
-  };
-
-  boxChecked = name => () => {
-    this.setState(state => {
-      state.selectedFunctions[name] = !state.selectedFunctions[name];
-      return state;
-    }, this.resetErrorMessage);
-  };
-
-  displayFunctions = () => {
-    return this.state.sourceFunctionList.map(sourceFunction => {
-      let name = sourceFunction.functionName;
-      let comment = sourceFunction.comment;
-      return (
-        <div key={name}>
-          <input
-            style={styles.largerCheckbox}
-            type="checkbox"
-            disabled={comment.length === 0}
-            name={name}
-            checked={this.state.selectedFunctions[name] || false}
-            onChange={this.boxChecked(name)}
-          />
-          <span>{name}</span>
-          <br />
-          {comment.length === 0 && (
-            <p style={styles.alert}>{i18n.libraryExportNoCommentError()}</p>
-          )}
-          <pre style={styles.textInput}>{comment}</pre>
-        </div>
-      );
-    });
-  };
-
-  displayAlert = () => {
-    let errorMessage;
-    if (this.state.publishState === PublishState.INVALID_INPUT) {
-      errorMessage = i18n.libraryPublishInvalid();
-    }
-    if (this.state.publishState === PublishState.ERROR_PUBLISH) {
-      errorMessage = i18n.libraryPublishFail();
-    }
-    return (
-      <div>
-        <p style={styles.alert}>{errorMessage}</p>
-      </div>
-    );
-  };
-
   displayContent = () => {
     return (
-      <div>
-        <Heading2>{i18n.libraryName()}</Heading2>
-        {this.displayNameInput()}
-        <Heading2>{i18n.description()}</Heading2>
-        {this.displayDescription()}
-        <Heading2>{i18n.catProcedures()}</Heading2>
-        {this.displayFunctions()}
-        <Button
-          style={{marginLeft: 0, marginTop: 20}}
-          onClick={this.publish}
-          text={i18n.publish()}
-        />
-        {this.displayAlert()}
-      </div>
+      <LibraryPublisher
+        onPublishSuccess={libraryName =>
+          this.setState({
+            dialogState: DialogState.PUBLISHED,
+            libraryName: libraryName
+          })
+        }
+        libraryDetails={this.state.libraryDetails}
+        clientApi={this.props.clientApi}
+      />
     );
   };
 
   render() {
     let bodyContent;
-    switch (this.state.publishState) {
-      case PublishState.LOADING:
+    switch (this.state.dialogState) {
+      case DialogState.LOADING:
         bodyContent = this.displayLoadingState();
         break;
-      case PublishState.PUBLISHED:
+      case DialogState.PUBLISHED:
         bodyContent = this.displaySuccess();
         break;
-      case PublishState.CODE_ERROR:
+      case DialogState.CODE_ERROR:
         bodyContent = this.displayError(i18n.libraryCodeError());
         break;
-      case PublishState.NO_FUNCTIONS:
+      case DialogState.NO_FUNCTIONS:
         bodyContent = this.displayError(i18n.libraryNoFunctonsError());
         break;
       default:
